Fixed the protocol buffers services

+++ back-end/go/internal/grpc/services/impl/discounted-date_create.go
@@ -28,7 +28,7 @@ func (d *DiscountedDateServiceServer) CreateDiscountedDate(ctx context.Context,
             "The description field is required and must be set to a non-empty value")
     }
 
-    if float32(request.DiscountedDate.DiscountPct) <= 0 {
+    if request.DiscountedDate.DiscountPct <= 0 {
         return nil, status.Error(codes.InvalidArgument,
             "The discount_pct field is required and must be set to a value greater than 0")
     }
